Move Intent to types
