+++ src/web/app/pages-instructor/instructor-course-details-page/instructor-course-details-page.component.ts
@@ -18,7 +18,7 @@ import {
   Students,
 } from '../../../types/api-output';
 import { ErrorMessageOutput } from '../../error-message-output';
-import { Intent } from '../../Intent';
+import { Intent } from '../../../types/Intent';
 import { StudentListSectionData, StudentListStudentData } from '../student-list/student-list-section-data';
 
 interface CourseStats {
