+++ src/web/app/pages-instructor/instructor-session-result-page/instructor-session-result-page.component.ts
@@ -9,7 +9,7 @@ import {
   SessionResults, Student, Students,
 } from '../../../types/api-output';
 import { ErrorMessageOutput } from '../../error-message-output';
-import { Intent } from '../../Intent';
+import { Intent } from '../../../types/Intent';
 import { InstructorSessionResultSectionType } from './instructor-session-result-section-type.enum';
 
 /**
