+++ src/web/app/pages-session/session-submission-page/session-submission-page.component.ts
@@ -29,7 +29,7 @@ import {
   QuestionSubmissionFormModel,
 } from '../../components/question-submission-form/question-submission-form-model';
 import { ErrorMessageOutput } from '../../error-message-output';
-import { Intent } from '../../Intent';
+import { Intent } from '../../../types/Intent';
 import {
   FeedbackSessionClosedModalComponent,
 } from './feedback-session-closed-modal/feedback-session-closed-modal.component';
