+++ src/web/app/pages.module.ts
@@ -6,7 +6,7 @@ import { ErrorReportModule } from './components/error-report/error-report.module
 import { LoaderBarModule } from './components/loader-bar/loader-bar.module';
 import { LoadingSpinnerComponent } from './components/loading-spinner/loading-spinner.component';
 import { StatusMessageModule } from './components/status-message/status-message.module';
-import { Intent } from './Intent';
+import { Intent } from '../types/Intent';
 import { PageNotFoundModule } from './page-not-found/page-not-found.module';
 import { ClickOutsideDirective, PageComponent } from './page.component';
 import { AdminPageComponent } from './pages-admin/admin-page.component';
